Refactoring: Moved dummy cart generator out, cleaned up the code
