+++ PromotionEngineTester/Domain/PromotionLogic/PromotionType1Tests.cs
@@ -1,11 +1,7 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using PromotionEngine.Domain.PromotionLogic;
-using PromotionEngineTester.Domain.PromotionLogic;
-using System;
+using PromotionEngineTester.Domain;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace PromotionEngine.Domain.PromotionLogic.Tests
 {
